pyproject.toml
[tool.poetry]
name = "rubikrypt"
version = "0.1.0"
description = ""
authors = ["Scott Odle <scott@sjodle.com>"]

[tool.poetry.dependencies]
python = "^3.10"
docopt = "^0.6.2"
rubik-cube = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rubikrypt = "rubikrypt.main:main"
